// sdk/src/generated/app-scale/ir/command.ts
// automatically generated by the FlatBuffers compiler, do not modify

/* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-explicit-any, @typescript-eslint/no-non-null-assertion */

import { AppendChild } from '../../app-scale/ir/append-child.js';
import { CreateNode } from '../../app-scale/ir/create-node.js';
import { InsertBefore } from '../../app-scale/ir/insert-before.js';
import { RemoveChild } from '../../app-scale/ir/remove-child.js';
import { SetRoot } from '../../app-scale/ir/set-root.js';
import { UpdateProps } from '../../app-scale/ir/update-props.js';
import { UpdateStyle } from '../../app-scale/ir/update-style.js';


export enum Command {
  NONE = 0,
  CreateNode = 1,
  UpdateProps = 2,
  UpdateStyle = 3,
  AppendChild = 4,
  InsertBefore = 5,
  RemoveChild = 6,
  SetRoot = 7
}

export function unionToCommand(
  type: Command,
  accessor: (obj:AppendChild|CreateNode|InsertBefore|RemoveChild|SetRoot|UpdateProps|UpdateStyle) => AppendChild|CreateNode|InsertBefore|RemoveChild|SetRoot|UpdateProps|UpdateStyle|null
): AppendChild|CreateNode|InsertBefore|RemoveChild|SetRoot|UpdateProps|UpdateStyle|null {
  switch(Command[type]) {
    case 'NONE': return null; 
    case 'CreateNode': return accessor(new CreateNode())! as CreateNode;
    case 'UpdateProps': return accessor(new UpdateProps())! as UpdateProps;
    case 'UpdateStyle': return accessor(new UpdateStyle())! as UpdateStyle;
    case 'AppendChild': return accessor(new AppendChild())! as AppendChild;
    case 'InsertBefore': return accessor(new InsertBefore())! as InsertBefore;
    case 'RemoveChild': return accessor(new RemoveChild())! as RemoveChild;
    case 'SetRoot': return accessor(new SetRoot())! as SetRoot;
    default: return null;
  }
}

export function unionListToCommand(
  type: Command, 
  accessor: (index: number, obj:AppendChild|CreateNode|InsertBefore|RemoveChild|SetRoot|UpdateProps|UpdateStyle) => AppendChild|CreateNode|InsertBefore|RemoveChild|SetRoot|UpdateProps|UpdateStyle|null, 
  index: number
): AppendChild|CreateNode|InsertBefore|RemoveChild|SetRoot|UpdateProps|UpdateStyle|null {
  switch(Command[type]) {
    case 'NONE': return null; 
    case 'CreateNode': return accessor(index, new CreateNode())! as CreateNode;
    case 'UpdateProps': return accessor(index, new UpdateProps())! as UpdateProps;
    case 'UpdateStyle': return accessor(index, new UpdateStyle())! as UpdateStyle;
    case 'AppendChild': return accessor(index, new AppendChild())! as AppendChild;
    case 'InsertBefore': return accessor(index, new InsertBefore())! as InsertBefore;
    case 'RemoveChild': return accessor(index, new RemoveChild())! as RemoveChild;
    case 'SetRoot': return accessor(index, new SetRoot())! as SetRoot;
    default: return null;
  }
}